+++ apps/web/src/components/GetStarted/GetNoticed.tsx
@@ -2,7 +2,7 @@ import {
   sectionContainer,
   titleClasses,
   GET_NOTICED_SECTION_ID,
-} from '../../../app/(base-org)/get-started/page';
+} from '../../../app/(base-org)/getstarted/page';
 import ResourceGrid from './ResourceGrid/ResourceGrid';
 import { ResourceSectionType } from './resourceTypes';
