+++ apps/web/src/components/GetStarted/StartBuilding.tsx
@@ -3,7 +3,7 @@ import {
   sectionContainer,
   titleClasses,
   START_BUILDING_SECTION_ID,
-} from '../../../app/(base-org)/get-started/page';
+} from '../../../app/(base-org)/getstarted/page';
 import { ButtonWithLinkAndEventLogging } from '../Button/ButtonWithLinkAndEventLogging';
 import ResourceGrid from './ResourceGrid/ResourceGrid';
 import { ResourceSectionType } from './resourceTypes';
